fine grain user permissions

+++ src/blueprints/bp_moderation.py
@@ -150,7 +150,7 @@ async def make_report_pagination(mod_status: ModStatus, board_shortnames: list[s
 @bp.get('/reports/closed/<int:page_num>')
 @login_required
 @require_is_active
-@require_permissions(set([Permissions.report_read]))
+@require_permissions([Permissions.report_read])
 async def reports_closed(page_num: int=0):
     page_size = 20
     reports = await get_reports_f(mod_status=ModStatus.closed, board_shortnames=board_shortnames, page_num=page_num, page_size=page_size)
@@ -171,7 +171,7 @@ async def reports_closed(page_num: int=0):
 @bp.get('/reports/open/<int:page_num>')
 @login_required
 @require_is_active
-@require_permissions(set([Permissions.report_read]))
+@require_permissions([Permissions.report_read])
 async def reports_open(page_num: int=0):
     page_size=20
     reports = await get_reports_f(mod_status=ModStatus.open, board_shortnames=board_shortnames, page_num=page_num, page_size=page_size)
@@ -202,6 +202,9 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
 
     match action:
         case 'report_delete':
+            if not current_user.has_permissions([Permissions.report_delete]):
+                return f'Need permissions for {Permissions.report_delete.name}'
+
             report = await delete_report_if_exists(report_parent_id)
             flash_msg = f'Report was already deleted.'
             if report:
@@ -210,6 +213,9 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
             return flash_msg
 
         case 'post_delete':
+            if not current_user.has_permissions([Permissions.post_delete]):
+                return f'Need permissions for {Permissions.post_delete.name}'
+            
             # Note: do not delete the report here. It is still needed to filter outgoing posts from full text search.
             post, result = await move_post_to_delete_table(report.board_shortname, report.num)
             if result == 0:
@@ -226,24 +232,36 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
             flash_msg += ' Deleted thumbnail.' if prev_del else ' Did not delete thumbnail.'
 
         case 'media_delete':
+            if not current_user.has_permissions([Permissions.media_delete]):
+                return f'Need permissions for {Permissions.media_delete.name}'
+            
             post = await get_post(report.board_shortname, report.num)
             full_del, prev_del = post_files_delete(post)
             flash_msg += ' Deleted full media.' if full_del else ' Did not delete full media.'
             flash_msg += ' Deleted thumbnail.' if prev_del else ' Did not delete thumbnail.'
 
         case 'media_hide':
+            if not current_user.has_permissions([Permissions.media_hide]):
+                return f'Need permissions for {Permissions.media_hide.name}'
+            
             post = await get_post(report.board_shortname, report.num)
             full_hid, prev_hid = post_files_hide(post)
             flash_msg += ' Hid full media.' if full_hid else ' Did not hide full media.'
             flash_msg += ' Hid thumbnail.' if prev_hid else ' Did not hide thumbnail.'
 
         case 'media_show':
+            if not current_user.has_permissions([Permissions.media_show]):
+                return f'Need permissions for {Permissions.media_show.name}'
+            
             post = await get_post(report.board_shortname, report.num)
             full_sho, prev_sho = post_files_show(post)
             flash_msg += ' Showing full media.' if full_sho else ' Did not reveal full media.'
             flash_msg += ' Showing thumbnail.' if prev_sho else ' Did not reveal thumbnail.'
 
         case 'post_show':
+            if not current_user.has_permissions([Permissions.post_show]):
+                return f'Need permissions for {Permissions.post_show.name}'
+            
             report = await edit_report_if_exists(report_parent_id, public_access=PublicAccess.visible)
             if report:
                 await fc.delete_post(report['board_shortname'], report['num'], report['op'])
@@ -256,6 +274,9 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
                 flash_msg += ' Showing thumbnail.' if prev_sho else ' Did not reveal thumbnail.'
 
         case 'post_hide':
+            if not current_user.has_permissions([Permissions.post_hide]):
+                return f'Need permissions for {Permissions.post_hide.name}'
+
             report = await edit_report_if_exists(report_parent_id, public_access=PublicAccess.hidden)
             if report:
                 await fc.insert_post(report['board_shortname'], report['num'], report['op'])
@@ -268,14 +289,23 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
                 flash_msg += ' Hid thumbnail.' if prev_hid else ' Did not hide thumbnail.'
 
         case 'report_close':
+            if not current_user.has_permissions([Permissions.report_close]):
+                return f'Need permissions for {Permissions.report_close.name}'
+
             report = await edit_report_if_exists(report_parent_id, mod_status=ModStatus.closed)
             flash_msg = 'Report moved to closed reports.'
 
         case 'report_open':
+            if not current_user.has_permissions([Permissions.report_open]):
+                return f'Need permissions for {Permissions.report_open.name}'
+
             report = await edit_report_if_exists(report_parent_id, mod_status=ModStatus.open)
             flash_msg = 'Report moved to opened reports.'
 
         case 'report_save_notes':
+            if not current_user.has_permissions([Permissions.report_save_notes]):
+                return f'Need permissions for {Permissions.report_save_notes.name}'
+
             # falsey mod_notes are valid
             await edit_report_if_exists(report_parent_id, mod_notes=mod_notes)
             flash_msg = f'Report had their moderation notes saved.'
@@ -289,7 +319,7 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
 @bp.route('/reports/<int:report_parent_id>/<string:action>', methods=['POST'])
 @login_required
 @require_is_active
-@require_permissions(set([Permissions.report_read]))
+@require_permissions([Permissions.report_read])
 async def reports_action(report_parent_id: int, action: str):
     form = (await request.form)
     redirect_endpoint = form.get('endpoint', 'bp_moderation.reports_open')
@@ -304,7 +334,7 @@ async def reports_action(report_parent_id: int, action: str):
 @bp.route('/reports/bulk/<string:action>', methods=['POST'])
 @login_required
 @require_is_active
-@require_permissions(set([Permissions.report_read]))
+@require_permissions([Permissions.report_read])
 async def reports_action_bulk(action: str):
     data = (await request.get_json())
